+++ changelog.md
@@ -20,6 +20,7 @@
 - removed recipe that converts Tattered Armor to Cloths
 - removed recipe that converts Cerberus Fangs to Bone Meal
 - removed check for natural spawns in Undead Army requirement
+- removed advancements for opening 16th Treasure Bag
 - made Bleeding immunity factor based on armor value and armor toughness (suggested by @Qee1y)
 - made blood particles darker the longer they last
 - made Giant launch attack weaker when player has knockback resistance